# api/swagger-api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/session": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Создание или обновление сессии",
        "description": "Если кука отсутствует или сессия истекла — создает новую (201). Если кука есть и сессия активна — обновляет её TTL в Redis (200).",
        "operationId": "session",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "description": "ID сессии из Cookie",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "3f8a2c1d9e4b7f0a5c6d2e8b1a3f9c7d"
          }
        ],
        "responses": {
          "201": {
            "description": "Сессия создана (первый визит)",
            "headers": {
              "Set-Cookie": {
                "description": "X-Session-Id={sid}; HttpOnly; Path=/; Max-Age={APP_USER_SESSION_TTL}",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "Сессия обновлена (повторный визит)",
            "headers": {
              "Set-Cookie": {
                "description": "X-Session-Id={sid}; HttpOnly; Path=/; Max-Age={APP_USER_SESSION_TTL}",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Выход из аккаунта",
        "description": "Удаляет текущую сессию и заставляет клиент удалить Cookie X-Session-Id (Max-Age=0).",
        "operationId": "logout",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "description": "ID сессии из Cookie",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешный выход",
            "headers": {
              "Set-Cookie": {
                "description": "Удаляет куку (Max-Age=0)",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Сессия не найдена или не авторизована"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Аутентификация пользователя",
        "description": "Проверяет credentials. При успехе привязывает сессию к пользователю и возвращает Cookie X-Session-Id.",
        "operationId": "login",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "description": "ID сессии из Cookie",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "3f8a2c1d9e4b7f0a5c6d2e8b1a3f9c7d"
          }
        ],
        "requestBody": {
          "description": "Данные пользователя",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "username": "j0hnd0e42",
                "password": "svp4_dvp4_str0ng_passw0rd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Успешная аутентификация",
            "headers": {
              "Set-Cookie": {
                "description": "Устанавливает X-Session-Id куку",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Неверный логин или пароль",
            "headers": {
              "Set-Cookie": {
                "description": "Обновляет TTL существующей куки",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "example": {
                  "message": "invalid credentials"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "description": "Данные для входа в систему",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя (username)",
            "example": "j0hnd0e42"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Пароль пользователя",
            "example": "svp4_dvp4_str0ng_passw0rd"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
...
    }
  }
}